Login - fix loading issue and update structure
